PCategoria borra

+++ PCategoria/PCategoria/MainWindow.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Data;
 using Gtk;
 using System.Collections;
 
@@ -26,25 +27,50 @@ public partial class MainWindow: Gtk.Window
 			QueryResult queryresult = PersisterHelper.Get("select * from categoria");
 			TreeViewHelper.Fill (treeview1, queryresult);
 		};
+
 		deleteAction.Activated += delegate {
-		TreeIter treeIter;
-		treeView.Selection.GetSelected (out treeIter);
-		IList row =(IList)treeView.Model.GetValue(treeIter,0);
-
-		string deleteSql = string.Format("delete from categoria where id={0}", id);
-		Console.WriteLine ("deleteSql={0}", deleteSql);
-
-
-
-		IDbCommand dbCommand = dbConnection.CreateCommand ();
-		dbCommand.CommandText = deleteSql;
-
-		dbCommand.ExecuteNonQuery ();
-		Destroy();
+			object id=TreeViewHelper.GetId(treeview1);
+			Console.WriteLine("click ondeleteAction id={0}", id);
+			delete(id);
 		};
-	}
-	
 
+		treeview1.Selection.Changed += delegate {
+			object id = TreeViewHelper.GetId (treeview1);
+			deleteAction.Sensitive = TreeViewHelper.IsSelected (treeview1);
+		};
+		deleteAction.Sensitive = false;
+	}
+
+		private void delete (object id){
+			if(ConfirmDelete(this)){
+				Console.WriteLine("Dice que eliminar si");
+				//Creamos un string que contenga comando para borrar
+				IDbCommand dbCommand = App.Instance.DbConnection.CreateCommand();
+				string borrar = string.Format("delete from categoria where id={0}", id);
+				dbCommand.CommandText = borrar;
+				dbCommand.ExecuteNonQuery();
+				fillTreeView (); //Actualiza sólo
+			};
+		}
+
+		public bool ConfirmDelete (Window window){
+			MessageDialog messageDialog = new MessageDialog(
+				window,
+				DialogFlags.DestroyWithParent,
+				MessageType.Question,
+				ButtonsType.YesNo,
+				"Quieres eliminar el elemento seleccionado?"
+				);
+			messageDialog.Title = Title;
+			ResponseType response = (ResponseType)messageDialog.Run ();
+			messageDialog.Destroy ();
+			return response == ResponseType.Yes;				
+		}
+	
+	private void fillTreeView() {
+		QueryResult queryResult = PersisterHelper.Get ("select * from categoria");
+		TreeViewHelper.Fill (treeview1, queryResult);
+	}
 	protected void OnDeleteEvent (object sender, DeleteEventArgs a)
 	{
 		Application.Quit ();
